Separando as rotas em um arquivo diferente de app.js
